# docs/v3-migration/openapi-v3.json
{
  "components": {
    "schemas": {
...
      "AdminLogin": {
        "properties": {
          "invite_code": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Invite Code",
            "type": "string"
          },
          "password": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Password",
            "type": "string"
          }
        },
        "required": [
          "password",
          "invite_code"
        ],
        "title": "AdminLogin",
        "type": "object"
      },
      "AdminLoginOut": {
        "properties": {
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "token": {
            "title": "Token",
            "type": "string"
          }
        },
        "required": [
          "token",
          "expires_at"
        ],
        "title": "AdminLoginOut",
        "type": "object"
      },
...
      "CustomerLegacyClaim": {
        "properties": {
          "device_label": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Label"
          },
          "display_name": {
            "default": "\u5973\u670b\u53cb",
            "maxLength": 50,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "invite_code": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Invite Code",
            "type": "string"
          },
          "legacy_customer_id": {
            "maxLength": 100,
            "minLength": 3,
            "title": "Legacy Customer Id",
            "type": "string"
          }
        },
        "required": [
          "invite_code",
          "legacy_customer_id"
        ],
        "title": "CustomerLegacyClaim",
        "type": "object"
      },
      "CustomerRecovery": {
        "description": "Recover an already claimed legacy identity or claim it for the first time.",
        "properties": {
          "device_label": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Label"
          },
          "display_name": {
            "default": "\u5973\u670b\u53cb",
            "maxLength": 50,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "invite_code": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Invite Code",
            "type": "string"
          },
          "legacy_customer_id": {
            "maxLength": 100,
            "minLength": 3,
            "title": "Legacy Customer Id",
            "type": "string"
          }
        },
        "required": [
          "invite_code",
          "legacy_customer_id"
        ],
        "title": "CustomerRecovery",
        "type": "object"
      },
      "CustomerSessionCreate": {
        "properties": {
          "device_label": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Label"
          },
          "display_name": {
            "default": "\u5973\u670b\u53cb",
            "maxLength": 50,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "invite_code": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Invite Code",
            "type": "string"
          }
        },
        "required": [
          "invite_code"
        ],
        "title": "CustomerSessionCreate",
        "type": "object"
      },
      "CustomerSessionOut": {
        "properties": {
          "customer_id": {
            "title": "Customer Id",
            "type": "string"
          },
          "customer_token": {
            "title": "Customer Token",
            "type": "string"
          },
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          }
        },
        "required": [
          "customer_id",
          "customer_token",
          "display_name"
        ],
        "title": "CustomerSessionOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "WeChatSessionCreate": {
        "description": "One-time WeChat login code plus optional first-login invite proof.",
        "properties": {
          "code": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Code",
            "type": "string"
          },
          "device_label": {
            "anyOf": [
              {
                "maxLength": 100,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": "\u5fae\u4fe1\u5c0f\u7a0b\u5e8f",
            "title": "Device Label"
          },
          "display_name": {
            "default": "\u5973\u670b\u53cb",
            "maxLength": 50,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "invite_code": {
            "default": "",
            "maxLength": 100,
            "title": "Invite Code",
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "title": "WeChatSessionCreate",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "\u60c5\u4fa3\u667a\u80fd\u53a8\u623f\u7ba1\u5bb6 API",
    "version": "3.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/admin/login": {
      "post": {
        "description": "Authenticate the database-owned administrator and retain the API contract.",
        "operationId": "admin_login_api_admin_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminLoginOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Admin Login"
      }
    },
...
    "/api/customers/claim-legacy": {
      "post": {
        "description": "Retain the original one-time legacy claim contract for old clients.",
        "operationId": "claim_legacy_customer_api_customers_claim_legacy_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerLegacyClaim"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerSessionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Claim Legacy Customer"
      }
    },
    "/api/customers/recover": {
      "post": {
        "description": "Recover a stable legacy identity and rotate any bearer that may be lost.",
        "operationId": "recover_customer_session_api_customers_recover_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerRecovery"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerSessionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recover Customer Session"
      }
    },
    "/api/customers/refresh": {
      "post": {
        "description": "Rotate the caller's current authenticated customer session.",
        "operationId": "refresh_customer_session_api_customers_refresh_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerSessionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Customer Session"
      }
    },
    "/api/customers/revoke": {
      "post": {
        "description": "Revoke only the current customer bearer and preserve all owned history.",
        "operationId": "revoke_customer_session_api_customers_revoke_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Customer Session"
      }
    },
    "/api/customers/session": {
      "post": {
        "description": "Create a rate-limited customer device session.",
        "operationId": "create_customer_session_api_customers_session_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomerSessionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerSessionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Customer Session"
      }
    },
    "/api/customers/wechat-session": {
      "post": {
        "description": "Exchange WeChat code, binding an authenticated legacy user in place.",
        "operationId": "create_wechat_customer_session_api_customers_wechat_session_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeChatSessionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerSessionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Wechat Customer Session"
      }
    },
...
  }
}